Moves plugin-specific common code to .plugins.common package.
