+++ jib-core/src/main/java/com/google/cloud/tools/jib/plugins/common/BuildStepsExecutionException.java
@@ -14,16 +14,14 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.frontend;
+package com.google.cloud.tools.jib.plugins.common;
 
-/** Thrown when main class inference fails. */
-public class MainClassInferenceException extends Exception {
+import com.google.cloud.tools.jib.builder.BuildSteps;
 
-  MainClassInferenceException(String message) {
-    super(message);
-  }
+/** Wraps an exception that happens during {@link BuildSteps#run}. */
+public class BuildStepsExecutionException extends Exception {
 
-  MainClassInferenceException(String message, Throwable cause) {
+  BuildStepsExecutionException(String message, Throwable cause) {
     super(message, cause);
   }
 }
