+++ jib-core/src/main/java/com/google/cloud/tools/jib/plugins/common/MainClassResolver.java
@@ -14,8 +14,10 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.frontend;
+package com.google.cloud.tools.jib.plugins.common;
 
+import com.google.cloud.tools.jib.frontend.MainClassFinder;
+import com.google.cloud.tools.jib.frontend.MainClassFinder.Result;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Splitter;
@@ -108,7 +110,7 @@ public class MainClassResolver {
                 + projectProperties.getJarPluginName()
                 + "; attempting to infer main class.");
 
-    MainClassFinder.Result mainClassFinderResult =
+    Result mainClassFinderResult =
         new MainClassFinder(
                 projectProperties.getClassesLayerEntry().getSourceFiles(),
                 projectProperties.getLogger())
