+++ jib-core/src/main/java/com/google/cloud/tools/jib/plugins/common/SystemPropertyValidator.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.frontend;
+package com.google.cloud.tools.jib.plugins.common;
 
 import java.util.function.Function;
