+++ jib-core/src/test/java/com/google/cloud/tools/jib/plugins/common/HelpfulSuggestionsTest.java
@@ -14,8 +14,9 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.frontend;
+package com.google.cloud.tools.jib.plugins.common;
 
+import com.google.cloud.tools.jib.plugins.common.HelpfulSuggestions;
 import java.nio.file.Paths;
 import org.junit.Assert;
 import org.junit.Test;
