+++ jib-core/src/test/java/com/google/cloud/tools/jib/plugins/common/SystemPropertyValidatorTest.java
@@ -14,8 +14,9 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.frontend;
+package com.google.cloud.tools.jib.plugins.common;
 
+import com.google.cloud.tools.jib.plugins.common.SystemPropertyValidator;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
